Clean up code, refactoring and change load, add and delete on async realizations
