+++ AccountantTool/AccountantTool.csproj
@@ -178,10 +178,6 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="View\AddAccountantRecordWindow.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
     <Page Include="View\MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
@@ -233,17 +229,6 @@
     <Compile Include="Controls\RequisitesControl.Designer.cs">
       <DependentUpon>RequisitesControl.cs</DependentUpon>
     </Compile>
-    <Compile Include="Converters\ListItemToIndexConverter.cs" />
-    <Compile Include="Data\AccountantDbContext.cs" />
-    <Compile Include="Data\AccountantDbInitializer.cs" />
-    <Compile Include="Data\Configs\AccountantRecordConfig.cs" />
-    <Compile Include="Data\Configs\AdditionalInfoConfig.cs" />
-    <Compile Include="Data\Configs\CompanyConfig.cs" />
-    <Compile Include="Data\Configs\ContactPersonConfig.cs" />
-    <Compile Include="Data\Configs\ContractConfig.cs" />
-    <Compile Include="Data\Configs\LicenseConfig.cs" />
-    <Compile Include="Data\Configs\ProductConfig.cs" />
-    <Compile Include="Data\Configs\RequisitesConfig.cs" />
     <Compile Include="Helpers\EditableListView.cs">
       <SubType>Component</SubType>
     </Compile>
@@ -280,15 +265,11 @@
     <Compile Include="ReoGrid\DataFormatter\LicenseDataFormatter.cs" />
     <Compile Include="ReoGrid\DataFormatter\ProductsDataFormatter.cs" />
     <Compile Include="ReoGrid\DataFormatter\RequisitesDataFormatter.cs" />
-    <Compile Include="ViewModel\AddAccountantRecordWindowViewModel.cs" />
-    <Compile Include="ViewModel\MainComponents\AsyncDelegateCommand.cs" />
-    <Compile Include="ViewModel\MainComponents\EventRaiser.cs" />
-    <Compile Include="ViewModel\MainComponents\RelayCommand.cs" />
-    <Compile Include="ViewModel\MainComponents\ViewModelBase.cs" />
+    <Compile Include="Common\AsyncDelegateCommand.cs" />
+    <Compile Include="Common\EventRaiser.cs" />
+    <Compile Include="Common\RelayCommand.cs" />
+    <Compile Include="Common\ViewModelBase.cs" />
     <Compile Include="ViewModel\MainWindowViewModel.cs" />
-    <Compile Include="View\AddAccountantRecordWindow.xaml.cs">
-      <DependentUpon>AddAccountantRecordWindow.xaml</DependentUpon>
-    </Compile>
     <Compile Include="View\MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
