--- AccountantTool/Converters/ListItemToIndexConverter.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Globalization;
-using System.Windows.Controls;
-using System.Windows.Data;
-
-namespace AccountantTool.Converters
-{
-    public class ListItemToIndexConverter : IValueConverter
-    {
-        public object Convert(object value, Type TargetType, object parameter, CultureInfo culture)
-        {
-            var item = value as ListViewItem;
-            var listView = ItemsControl.ItemsControlFromItemContainer(item) as ListView;
-            int index = listView.ItemContainerGenerator.IndexFromContainer(item) + 1;
-            return index.ToString();
-        }
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
