--- AccountantTool/Data/Configs/AdditionalInfoConfig.cs
@@ -1,14 +1,0 @@
-﻿using System.Data.Entity.ModelConfiguration;
-using AccountantTool.Model;
-
-namespace AccountantTool.Data.Configs
-{
-    public class AdditionalInfoConfig : EntityTypeConfiguration<AdditionalInfo>
-    {
-        //TODO: Update config if it is need
-        public AdditionalInfoConfig()
-        {
-            
-        }
-    }
-}
