--- AccountantTool/Data/Configs/CompanyConfig.cs
@@ -1,14 +1,0 @@
-﻿using System.Data.Entity.ModelConfiguration;
-using AccountantTool.Model;
-
-namespace AccountantTool.Data.Configs
-{
-    public class CompanyConfig : EntityTypeConfiguration<Company>
-    {
-        //TODO: Update config if it is need
-        public CompanyConfig()
-        {
-            
-        }
-    }
-}
