--- AccountantTool/Data/Configs/ContactPersonConfig.cs
@@ -1,14 +1,0 @@
-﻿using System.Data.Entity.ModelConfiguration;
-using AccountantTool.Model;
-
-namespace AccountantTool.Data.Configs
-{
-    public class ContactPersonConfig : EntityTypeConfiguration<ContactPerson>
-    {
-        //TODO: Update config if it is need
-        public ContactPersonConfig()
-        {
-            
-        }
-    }
-}
