--- AccountantTool/Data/Configs/ContractConfig.cs
@@ -1,14 +1,0 @@
-﻿using System.Data.Entity.ModelConfiguration;
-using AccountantTool.Model;
-
-namespace AccountantTool.Data.Configs
-{
-    public class ContractConfig : EntityTypeConfiguration<Contract>
-    {
-        //TODO: Update config if it is need
-        public ContractConfig()
-        {
-            
-        }
-    }
-}
