--- AccountantTool/Data/Configs/LicenseConfig.cs
@@ -1,14 +1,0 @@
-﻿using System.Data.Entity.ModelConfiguration;
-using AccountantTool.Model;
-
-namespace AccountantTool.Data.Configs
-{
-    public class LicenseConfig : EntityTypeConfiguration<License>
-    {
-        //TODO: Update config if it is need
-        public LicenseConfig()
-        {
-            
-        }
-    }
-}
