--- AccountantTool/Data/Configs/ProductConfig.cs
@@ -1,14 +1,0 @@
-﻿using System.Data.Entity.ModelConfiguration;
-using AccountantTool.Model;
-
-namespace AccountantTool.Data.Configs
-{
-    public class ProductConfig : EntityTypeConfiguration<Product>
-    {
-        //TODO: Update config if it is need
-        public ProductConfig()
-        {
-            
-        }
-    }
-}
