--- AccountantTool/Data/Configs/RequisitesConfig.cs
@@ -1,14 +1,0 @@
-﻿using System.Data.Entity.ModelConfiguration;
-using AccountantTool.Model;
-
-namespace AccountantTool.Data.Configs
-{
-    public class RequisitesConfig : EntityTypeConfiguration<Requisites>
-    {
-        //TODO: Update config if it is need
-        public RequisitesConfig()
-        {
-            
-        }
-    }
-}
